+++ Backpack/RTLSupport/Classes/BPKRTLSupport.h
@@ -15,21 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#import <UIKit/UIKit.h>
 
-#ifndef __BACKPACK__
-    #define __BACKPACK__
-
-    #import "Color.h"
-    #import "Font.h"
-    #import "Radii.h"
-    #import "Shadow.h"
-    #import "Spacing.h"
-    #import "Gradient.h"
-
-    #import "Badge.h"
-    #import "Label.h"
-    #import "Panel.h"
-    #import "Icon.h"
-    #import "Button.h"
-#endif
-
+NS_ASSUME_NONNULL_BEGIN
+@interface BPKRTLSupport: NSObject
++ (BOOL)viewIsRTL:(UIView *)view;
++ (UIEdgeInsets)makeRTLAwareEdgeInsetsForView:(UIView *)view
+                                          top:(CGFloat)top
+                                      leading:(CGFloat)leading
+                                       bottom:(CGFloat)bottom
+                                     trailing:(CGFloat)trailing;
+@end
+NS_ASSUME_NONNULL_END