+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 [Unreleased changes](./UNRELEASED.md).
 
+## 0.13.0
+
+**Breaking:**
+
+- Backpack/Color
+  - Update Swift interface for Backpack colors. `BPKColor` is now `Backpack.Color` and the individual colors are class properties. `BPKColor.blue500()` becomes `Backpack.Color.blue500`.
+
+## 0.12.0
+
+**Added:**
+- Backpack/Icon
+  - Added new method for creating images for use as templates. `+ (UIImage *)templateIconNamed:name size:` in Objective-C and `makeTemplateIcon(name:size:)` in Swift.
+- Backpack/Button
+  - Introducing the Backpack button. Includes all Backpack button styles, support for leading and trailing icons as well as icon only buttons.
+
 ## 0.11.15
 
 **Added:**